pyproject.toml
[tool.poetry]
name = "natasha-bot"
version = "0.1.0"
description = ""
authors = ["Slava Kuryaev <92830706+necrosskull@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "natasha_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"
python-telegram-bot = {extras = ["job-queue"], version = "^20.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
